get all tests

+++ src/main/java/ca/mcgill/ecse321/gameshop/model/Payment.java
@@ -25,7 +25,7 @@ public class Payment
 
   //Payment Attributes
   private String billingAddress;
-  private int creditCardNb;
+  private long creditCardNb;
   private String expirationDate;
   private int cvc;
   @Id
@@ -45,7 +45,7 @@ public class Payment
   public Payment(){
 
   }
-  public Payment(String aBillingAddress, int aCreditCardNb, String aExpirationDate, int aCvc)
+  public Payment(String aBillingAddress, long aCreditCardNb, String aExpirationDate, int aCvc)
   {
     billingAddress = aBillingAddress;
     creditCardNb = aCreditCardNb;
@@ -54,13 +54,13 @@ public class Payment
     commands = new ArrayList<Command>();
 
   }
-  public Payment(String aBillingAddress, int aCreditCardNb, String aExpirationDate, int aCvc,  Customer aCustomer)
+  public Payment(String aBillingAddress, long aCreditCardNb, String aExpirationDate, int aCvc,  Customer aCustomer)
   {
     billingAddress = aBillingAddress;
     creditCardNb = aCreditCardNb;
     expirationDate = aExpirationDate;
     cvc = aCvc;
-   // paymentId = aPaymentId;
+    // paymentId = aPaymentId;
     boolean didAddCustomer = setCustomer(aCustomer);
     if (!didAddCustomer)
     {
@@ -81,7 +81,7 @@ public class Payment
     return wasSet;
   }
 
-  public boolean setCreditCardNb(int aCreditCardNb)
+  public boolean setCreditCardNb(long aCreditCardNb)
   {
     boolean wasSet = false;
     creditCardNb = aCreditCardNb;
@@ -118,7 +118,7 @@ public class Payment
     return billingAddress;
   }
 
-  public int getCreditCardNb()
+  public long getCreditCardNb()
   {
     return creditCardNb;
   }
